ai-yt-optimiser/requirements.txt
google-auth-oauthlib
google-api-python-client
youtube-transcript-api
openai
python-dotenv
customtkinter
